Fix errors where arguments should be double instead of int.
- We really should just get rid of set(int index, double value) in
geometry classes like Tuple2D

+++ IHMCSwingPlotting/src/us/ihmc/plotting/PlotterLegendPanel.java
@@ -60,7 +60,7 @@ public class PlotterLegendPanel extends JPanel implements ArtifactsChangedListen
          y = y + deltaY;
          g.setFont(f);
 
-         drawOrigin.set(artifactX, y);
+         drawOrigin.set((double) artifactX, (double) y);
          
          artifact.drawLegend(graphics2dAdapter, drawOrigin);
